cloud_functions/admin/requirements.txt
functions-framework==3.0.0
google-cloud-bigquery==2.34.4
google-cloud-secretmanager==2.12.4
